+++ pages/news.php
@@ -7,34 +7,34 @@ Debugger::enable(Debugger::DETECT, $config['folder_logs']);
 if (isset($URL['3']) and $URL['1'] == "news" and ($user['aclNews'] > 0) and $URL['2'] == "delete") { // DELETE
     mysqli_query($database, "UPDATE ".DB_PREFIX."news set deleted=1 where id='".$URL['3']."'");
     if (mysqli_affected_rows($database) == 1) {
-        auditTrail(5, 11, $URL['3']);
+        authorizedAccess(5, 11, $URL['3']);
         $latteParameters['message'] = $text['aktualitaodebrana'];
     } else {
         $latteParameters['message'] = $text['aktualitaneodebrana'];
     }
 } elseif (isset($_GET['newsdelete'])) {
     $latteParameters['message'] = $text['http401'];
-    unauthorizedAccess(5, 0, 0, $URL[3]);
+    unauthorizedAccess(5, 11, $URL[3]);
 }
 
 if (isset($URL['3']) and $URL['1'] == "news" and ($user['aclNews'] > 0) and $URL['2'] == "restore") { // DELETE
     mysqli_query($database, "UPDATE ".DB_PREFIX."news set deleted=0 where id='".$URL['3']."'");
     if (mysqli_affected_rows($database) == 1) {
-        auditTrail(5, 11, $URL['3']);
+        authorizedAccess(5, 11, $URL['3']);
         $latteParameters['message'] = $text['aktualitaobnovena'];
     } else {
         $latteParameters['message'] = $text['aktualitaneobnovena'];
     }
 } elseif (isset($_GET['newsdelete'])) {
     $latteParameters['message'] = $text['http401'];
-    unauthorizedAccess(5, 0, 0, $URL[3]);
+    unauthorizedAccess(5, 11, $URL[3]);
 }
 
 if ($URL['1'] == "news" and ($user['aclNews'] > 0) and isset($_POST['news_new'])) { // ADD
     if ($_POST['insertnews'] && !preg_match('/^[[:blank:]]*$/i', $_POST['nadpis']) && !preg_match('/^[[:blank:]]*$/i', $_POST['news_new']) && is_numeric($_POST['kategorie'])) {
         mysqli_query($database, "INSERT INTO ".DB_PREFIX."news ( datum, iduser, kategorie, nadpis, obsahMD, deleted) VALUES('".time()."','".$user['userId']."','".$_POST['kategorie']."','".$_POST['nadpis']."','".$_POST['news_new']."',0)");
         if (mysqli_affected_rows($database) == 1) {
-            auditTrail(5, 3, 0);
+            authorizedAccess(5, 3, 0);
             $latteParameters['message'] = $text['aktualitavlozena'];
             unreadRecords(5, 0);
         } else {