+++ symbol_search.php
@@ -5,7 +5,7 @@ Debugger::enable(Debugger::DETECT,$config['folder_logs']);
 latteDrawTemplate("header");
 
 $latteParameters['title'] = 'Symboly';
-	auditTrail(7, 14, 0);
+	authorizedAccess(7, 14, 0);
 	mainMenu ();
 	sparklets ('<a href="./persons.php">osoby</a> &raquo; <a href="newperson.php">přidat osobu</a>; <a href="symbols.php">nepřiřazené symboly</a>; <a href="symbol_search.php">vyhledat symbol</a>');
 	?>
@@ -26,7 +26,7 @@ $latteParameters['title'] = 'Symboly';
 				<option>8</option>
 				<option>9</option>
 				<option>10</option>
-			</datalist>		
+			</datalist>
 	    <div class="central_frame">
 	        <div class="input_frame">
 	        	<div class="input_text">